% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kappa.R
\name{calc_kappa}
\alias{calc_kappa}
\title{Calculate condition number kappa}
\usage{
calc_kappa(mod, data = NULL)
}
\arguments{
\item{mod}{a model object created by \code{lm, glm, lmer, glmer}}
}
\value{
the condition number
}
\description{
Calculates the condition number with the intercept included, following Belsley, Kuh and Welsch (1980).
}
\details{
Adapted from the \code{\link[languageR]{collin.fnc}} function in the \code{\link[languageR]{languageR}} package.
}
\examples{
\dontrun{
set.seed(1)
x1 <- rnorm(100)
x2 <- x1+.1*rnorm(100)
x3 <- x2 + .5*rnorm(100)
y  <- sample(0:1, 100, TRUE)
f  <- glm(y ~ x1 + x2 + x3, family = binomial, x = TRUE)
calc_kappa(f)
}
}
\references{
Belsley, D. A. and Kuh, E. and Welsch, R. E. (1980) Regression Diagnostics. Identifying Influential Data and Sources of Collinearity, Wiley Series in Probability and Mathematical Statistics, New York.
}
\author{
F.J. Tweedie; Jason Grafmiller
}
