% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_max_distances.R
\name{calc_max_distances}
\alias{calc_max_distances}
\title{Calculate the maximum distances for a set of possible constraints}
\usage{
calc_max_distances(n = 6:20)
}
\arguments{
\item{n}{vector of integers representing different numbers of constraints}
}
\value{
a dataframe
}
\description{
Calculate the maximum distances for a set of possible constraints
}
\examples{
\dontrun{
calc_max_distances(6:20)
}
}
