% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_VIF.R
\name{calc_VIF}
\alias{calc_VIF}
\title{Calculate variance inflation factors for regression model parameters}
\usage{
calc_VIF(mod)
}
\arguments{
\item{mod}{a model object created by \code{lm, glm, lmer, glmer}}
}
\value{
a vector of VIFs
}
\description{
Computes variance inflation factors from the covariance matrix of parameter estimates, using the method of Davis et al. (1986), which is based on the correlation matrix from the information matrix.
}
\details{
Adapted from \code{\link[rms]{vif}} in the \code{\link{rms}} package.
}
\examples{
\dontrun{
set.seed(1)
x1 <- rnorm(100)
x2 <- x1+.1*rnorm(100)
x3 <- x2 + .5*rnorm(100)
y  <- sample(0:1, 100, TRUE)
f  <- glm(y ~ x1 + x2 + x3, family = binomial)
calc_VIF(f)
}
}
\references{
Davis CE, Hyde JE, Bangdiwala SI, Nelson JJ: An example of dependencies among variables in a conditional logistic regression. In Modern Statistical Methods in Chronic Disease Epidemiology, Eds SH Moolgavkar and RL Prentice, pp. 140–147. New York: Wiley; 1986.
}
\author{
Frank Harrell; Jason Grafmiller
}
