% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coef_table.R
\name{create_coef_table}
\alias{create_coef_table}
\title{Create Table of Coefficients for VADIS analysis}
\usage{
create_coef_table(mod_list)
}
\arguments{
\item{mod_list}{A list of regression model objects.}

\item{path}{Path in which to save the output (as \code{.csv} file). If \code{NULL}, defaults to the current working directory. Set \code{path = FALSE} if you do not wish to save to file.}
}
\value{
A dataframe
}
\description{
Create Table of Coefficients for VADIS analysis
}
\details{
The function loops through a list of model objects, extracts the coefficient estimates, and compiles them in a single dataframe.
}
\examples{
\dontrun{
lm_fnc <- function(x) lm(Sepal.Length ~ Petal.Length + Petal.Width, data = x)
rm_list <- fit.vadis.RM(iris, split.by = "Species", fit.func = lm_fnc,
  path = FALSE)
summary(rm_list[[1]])

create_coef_table(rm_list, path = FALSE)
}
}
\author{
Jason Grafmiller
}
