% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mds_samples.R
\name{create_mds_samples}
\alias{create_mds_samples}
\title{Create dataframe of coordinates for multidmensional scaling of posterior samples}
\usage{
create_mds_samples(
  mod_list,
  n_samples = 200L,
  k = 3,
  vadis_line = NULL,
  path = NULL
)
}
\arguments{
\item{mod_list}{list of regression model objects of class \code{brmsfit}.}

\item{n_samples}{integer indicating number of samples to pull}

\item{k}{= the maximum dimension of the space which the data are to be represented in the MDS; must be in ${1, 2, …, n-1}$ (see \code{\link[stats]{cmdscale}}).}

\item{vadis_line}{list containing the output of \code{vadis_line2}. If \code{NULL} (default), this will be run automatically}

\item{path}{filepath to which to save the output as an .csv file. If \code{NULL}, defaults to the current working directory. Set \code{path = FALSE} if you do not wish to save to file.}
}
\value{
a dataframe
}
\description{
Create dataframe of coordinates for multidmensional scaling of posterior samples
}
\details{
The function creates a dataframe containing MDS coordinates of distance matrices derived from posterior samples of coefficient estimates. Currently only works with \code{brmsfit} objects.
}
\examples{
\dontrun{
Coming soon...
}
}
\author{
Jason Grafmiller
}
