% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signif_table.R
\name{create_signif_table}
\alias{create_signif_table}
\title{Create Table of Significant Effects for VADIS analysis}
\usage{
create_signif_table(
  mod_list,
  alpha = 0.05,
  method = c("freq", "pd", "rope", "map")
)
}
\arguments{
\item{mod_list}{A list of regression model objects.}

\item{method}{string indicating which method to use for assessing significance from Bayesian models. See \code{\link[VADIS]{vadis_line1}} for details.}

\item{path}{Path in which to save the output (as \code{.csv} file). If \code{NULL}, defaults to the current working directory. Set \code{path = FALSE} if you do not wish to save to file.}
}
\value{
A dataframe
}
\description{
Create Table of Significant Effects for VADIS analysis
}
\details{
The function loops through a list of model objects, extracts the significance values of the parameter estimates, and compiles them in a single dataframe.
}
\examples{
\dontrun{
lm_fnc <- function(x) lm(Sepal.Length ~ Petal.Length + Petal.Width, data = x)
data_list <- split(iris, list(iris$Species), drop = T)
rm_list <- lapply(data_list, FUN = function(x) glm(Petal.Length ~ ., x))

create_coef_table(rm_list)
}
}
\author{
Jason Grafmiller
}
