% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_samples.R
\name{get_posterior_samples}
\alias{get_posterior_samples}
\title{Retrieve random samples of posterior coefficient estimates}
\usage{
get_posterior_samples(mod_list, n = 200L)
}
\arguments{
\item{mod_list}{A list of regression model objects of class \code{brmsfit}.}

\item{n}{Integer indicating number of samples to pull}
}
\value{
a list of dataframes
}
\description{
Retrieve random samples of posterior coefficient estimates
}
\details{
The function loops through a list of *M* model objects, randomly samples coefficient estimates from the posterior, and compiles those samples in a single dataframe. Output is a list of *M* dataframes each with *n* columns of sampled estimates. Currently only works with \code{brmsfit} objects.
}
\examples{
\dontrun{
Coming soon...
}
}
\author{
Jason Grafmiller
}
