% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosmerlem_test.R
\name{hosmerlem_test}
\alias{hosmerlem_test}
\title{Hosmer-Lemeshow goodness of fit test}
\usage{
hosmerlem_test(mod, k = 10)
}
\arguments{
\item{mod}{a model object created by \code{lm, glm, lmer, glmer}}

\item{k}{number of bins to use to calculate quantiles}
}
\value{
a dataframe
}
\description{
The Hosmer-Lemeshow test is a statistical test for goodness of fit for logistic regression models.
}
\examples{
\dontrun{
set.seed(43214)
x <- rnorm(100)
pr <- exp(x)/(1+exp(x))
y <- 1*(runif(100) < pr)
mod <- glm(y~x, family=binomial)
hosmerlem_test(mod, k = 10)
}
}
\references{

}
