% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Min-Max scaling}
\usage{
minmax(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector
}
\description{
function for scaling a numeric vector to between 0 and 1
}
\examples{
x <- rnorm(20)
minmax(x)
}
