% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hierarchy.R
\name{plot_hierarchy}
\alias{plot_hierarchy}
\title{Plot hierarchy of a constraint's levels across multiple models}
\usage{
plot_hierarchy(focal.predictors = NULL, mod_list, data = data_list, ncol = 3)
}
\arguments{
\item{ncol}{}
}
\value{

}
\description{
Plot hierarchy of a constraint's levels across multiple models
}
