% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.R
\name{stand}
\alias{stand}
\title{Standardize regression inputs by centering around the mean and dividing by 2 standard deviations}
\usage{
stand(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A numeric vector
}
\description{
Standardize regression inputs by centering around the mean and dividing by 2 standard deviations
}
\details{
Inspired by the \code{\link[arm]{standardize}} function in the \code{arm} package. Numeric predictors are rescaled by subtracting the mean  and dividing by 2 standard deviations. Binary predictors are rescaled to have a mean of 0 and a difference of 1 between their two levels. Non-numeric variables with more than two values are unchanged.
}
\examples{
\dontrun{
x <- rnorm(10, 5, 3)
stand(x)

y <- sample(c("A", "B"), 10, replace = T, prob = c(.8, .2))
stand(y)
}
}
\seealso{
\code{\link[arm]{standardize}}
}
\author{
Jason Grafmiller
}
