% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vadis_line1.R
\name{vadis_line1}
\alias{vadis_line1}
\title{Calculate the first line of evidence for the VADIS method}
\usage{
vadis_line1(
  mod_list,
  path = NULL,
  alpha = 0.05,
  method = c("freq", "pd", "rope", "map"),
  overwrite = c("no", "reload", "yes")
)
}
\arguments{
\item{mod_list}{A list of regression model objects.}

\item{path}{Path in which to save the output as an R data file (\code{.rds}). If \code{NULL}, defaults to the current working directory. Set \code{path = FALSE} if you do not wish to save to file.}

\item{alpha}{The significance threshold. Default is .05}

\item{method}{The method for calculating significance values. See details.}

\item{overwrite}{Should the function overwrite data to location in \code{path}? Default is \code{'no'}. If 'no', you will be asked to enter a new file location. Specify \code{path = 'reload'} to reload as existing file.}
}
\value{
A \code{list} of length 3.
\describe{
\item{\code{signif.table}}{A dataframe of \emph{P} predictors by \emph{M} models, containing a binary value indicating statistical significance (1 = significant) for each predictor in each model.}
\item{\code{distance.matrix}}{An \emph{M} by \emph{M} distance matrix of class \code{dist}, derived from \code{signif.table}. Values are squared Euclidean distances normalized by the number of predictors \emph{P}.}
\item{\code{similarity.scores}}{A dataframe of similarity scores derive from \code{distance.matrix}. See Szmrecsanyi et al. (2019) for details.}
}
}
\description{
Calculate the first line of evidence for the VADIS method
}
\details{
The function loops through a list of model objects, extracts the coefficient estimates, and compiles them in a single dataframe. There are four possible values for the \code{method} argument: \code{"freq"}, \code{"pd"}, \code{"rope"}, \code{"map"}. For frequentist (non-Bayesian) models, only the standard outputs of \code{glm} and \code{glmer} are used, i.e. \code{method = "freq"}.

For Bayesian models, all four methods are available. For \code{method = "freq"}, significance is determined based on the Highest Posterior Density Interval (HDI), which is determined as 1 - \code{alpha}. Significance is defined as whether the HDI contains 0. For \code{method = "pd"} ("probability of direction"), the p-value is defined as the proportion of the posterior distribution that is of the median’s sign. In other words, the p-value represents the proportion of the posterior distribution that is above/below 0, whichever is larger). For \code{method = "rope"}, the p-value is defined as the proportion of the entire posterior distribution that lies within the Region of Practical Equivalence (ROPE), which is defined here as \code{c(-0.1, 0.1)} (see Kruschke \& Liddell 2018; Makowski et al. 2019). For \code{method = "map"}, the p-value is defined as the density value at 0 divided by the density at the Maximum A Posteriori (MAP). See \code{\link[bayestestR]{p_direction}}, \code{\link[bayestestR]{p_rope}}, and \code{\link[bayestestR]{p_map}} for further details.
}
\examples{
\dontrun{
data_list <- split(particle_verbs_short, particle_verbs_short$Variety, drop = TRUE)

fmla <- Response ~ DirObjWordLength + DirObjDefiniteness + DirObjGivenness + DirObjConcreteness + DirObjThematicity + DirectionalPP + PrimeType + Semantics + Surprisal.P + Surprisal.V + Register

glm_func <- function(x) glm(fmla, data = x, family = binomial)

glm_list <- lapply(data_list, glm_func)
names(glm_list) <- names(data_list)

line1 <- vadis_line1(glm_list, path = FALSE)
}
}
\references{
Kruschke, John K. & Torrin M. Liddell. 2018. The Bayesian New Statistics: Hypothesis testing, estimation, meta-analysis, and power analysis from a Bayesian perspective. \emph{Psychonomic Bulletin & Review} 25(1). 178–206. doi: 10.3758/s13423-016-1221-4.

Makowski, Dominique, Mattan S. Ben-Shachar, S. H. Annabel Chen & Daniel Lüdecke. 2019. Indices of effect existence and significance in the Bayesian framework. \emph{Frontiers in Psychology}. 10. doi: 10.3389/fpsyg.2019.02767.

Szmrecsanyi, Benedikt, Jason Grafmiller & Laura Rosseel. 2019. Variation-Based Distance and Similarity Modeling: A Case Study in World Englishes. \emph{Frontiers in Artificial Intelligence} 2. \url{https://doi.org/10.3389/frai.2019.00023}.
}
\author{
Jason Grafmiller
}
