% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vadis_line2.R
\name{vadis_line2}
\alias{vadis_line2}
\title{Calculate the second line of evidence for the VADIS method}
\usage{
vadis_line2(mod_list, path = NULL)
}
\arguments{
\item{mod_list}{A list of regression model objects.}

\item{path}{Path in which to save the output as an R data file (\code{.rds}). If \code{NULL}, defaults to the current working directory. Set \code{path = FALSE} if you do not wish to save to file.}
}
\value{
A dataframe
}
\description{
Calculate the second line of evidence for the VADIS method
}
\details{
The function loops through a list of model objects, extracts the coefficient estimates, and compiles them in a single dataframe.
}
\examples{
\dontrun{
lm_fnc <- function(x) lm(Sepal.Length ~ Petal.Length + Petal.Width + Sepal.width, data = x)
rm_list <- fit.vadis.RM(iris, split.by = "Species", fit.func = lm_fnc,
  path = FALSE)
summary(rm_list[[1]])

line2 <- calc_line2(rm_list, path = FALSE)
}
}
\author{
Jason Grafmiller
}
