% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rank_table.R
\name{create_rank_table}
\alias{create_rank_table}
\title{Create Table of Variable Importance Rankings for VADIS analysis}
\usage{
create_rank_table(mod_list, conditional = FALSE)
}
\arguments{
\item{mod_list}{A list of random forest model objects, generated with \code{fit_vadis_rf()}. Currently supports objects of \code{\link[party]{RandomForest-class}}, \code{\link[ranger]{ranger}}, and \code{\link[randomForest]{randomForest}}}

\item{conditional}{logical. Should unconditional (default) or conditional permutation variable importance be computed. Only applies to \code{RandomForest-class} models from the \code{\link[party]{party}} package.}

\item{path}{Path in which to save the output (as \code{.csv} file). If \code{NULL}, defaults to the current working directory. Set \code{path = FALSE} if you do not wish to save to file.}
}
\value{
A dataframe
}
\description{
Create Table of Variable Importance Rankings for VADIS analysis
}
\details{
The function loops through a list of random forest objects, extracts the variable importance estimates, and compiles them in a single dataframe.
For forests fit with \code{ranger} or \code{randomForest}, the \code{importance} argument must be specified.
#'
}
\examples{
\dontrun{
fmla <- Type ~ PossrAnimacyBin + PossrWordC + PossmWordC + FinalSibilant +
  TypeTokenRatio + ProtoSemanticRelation + PossrExpType

rf_fnc <- function(x) ranger::ranger(fmla, data = x, importance = "permutation")

rf_list <- fit_models(brown_genitives, split.by = "Genre", fit.func = rf_fnc, path = FALSE)

create_rank_table(rf_list, path = FALSE)
}
}
\author{
Jason Grafmiller
}
