% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.R
\name{stand}
\alias{stand}
\title{Standardize regression inputs by centering around the mean and dividing by 2 standard deviations}
\usage{
stand(x, cols = NULL)
}
\arguments{
\item{x}{A vector, matrix, or dataframe}

\item{cols}{A character vector of column names, a numeric vector of column indices, or a formula}
}
\value{
A numeric vector, a matrix, or a dataframe
}
\description{
Standardize regression inputs by centering around the mean and dividing by 2 standard deviations
}
\details{
Inspired by the \code{\link[arm]{standardize}} function in the \code{arm} package. Numeric predictors are rescaled by subtracting the mean  and dividing by 2 standard deviations. Binary predictors are rescaled to have a mean of 0 and a difference of 1 between their two levels. Non-numeric variables with more than two values are unchanged.
}
\examples{
\dontrun{
x <- rnorm(10, 5, 3)
x
stand(x)

y <- sample(c("A", "B"), 10, replace = T, prob = c(.8, .2))
y
stand(y)

df <- data.frame(X = x, Y = y)
df
stand(df, 1:2)
stand(df, Y ~ X)
}
}
\seealso{
\code{\link[arm]{standardize}}
}
\author{
Jason Grafmiller
}
