% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Calculate summary statistics for list of models}
\usage{
summary_stats(mlist, dlist = NULL, response = NULL)
}
\arguments{
\item{mlist}{a list of model objects}

\item{dlist}{a list od dataframes}

\item{response}{the names of the response column in the data}
}
\value{
a vector
}
\description{
Calculate summary statistics for list of models
}
\examples{
\dontrun{
data_list <- vector("list")
for (i in 1:4){
  df <- data.frame(x1 = rnorm(100))
  df$x2 <- df$x1+.1*rnorm(100)
  df$x3 <- df$x2 + .5*rnorm(100)
  df$y <- rbinom(100, 1, 1/(1 + exp(-df$x1+df$x2+df$x3)))
 data_list[[i]] <- df
}
rm_list <- lapply(data_list, FUN = function(d) glm(y ~ ., data = d, family = binomial, x = T))
summary_stats(rm_list)
}
}
