% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vadis_line2.R
\name{vadis_line2}
\alias{vadis_line2}
\title{Calculate the second line of evidence for the VADIS method}
\usage{
vadis_line2(
  mod_list,
  path = NULL,
  weight = 1,
  scale = c("abs", "mean", "minmax", "none"),
  overwrite = c("no", "yes", "reload"),
  verbose = FALSE
)
}
\arguments{
\item{mod_list}{A list of regression model objects.}

\item{path}{Path in which to save the output as an R data file (\code{.rds}). If \code{NULL}, defaults to the current working directory. Set \code{path = FALSE} if you do not wish to save to file.}

\item{weight}{A numeric value indicating the size of the "effects" used for approximating the maximal reasonable distance. Default is 1.}

\item{scale}{How should the distance matrix be scaled? See details}

\item{overwrite}{Should the function overwrite data to location in \code{path}? Default is \code{'no'}, which will run the analysis if no file exists. If file in \code{path} exists, user with be prompted to set new path or allow file to be overwritten. Set to \code{'yes'} to automatically overwrite existing file, and \code{'reload'} to automatically reload existing file.}

\item{verbose}{Should messages be printed? Default is \code{FALSE}}
}
\value{
A \code{list} of length 3.
\describe{
\item{\code{coef.table}}{A dataframe of \emph{P} predictors by \emph{M} models, containing the pointwise estimated  coefficients (for \code{glm} and \code{glmer} models) or the mean posterior \beta estimates (for \code{brmsfit} models) for each predictor in each model.}
\item{\code{distance.matrix}}{An \emph{M} by \emph{M} distance matrix of class \code{dist}, derived from \code{coef.table}. Values are (normalized) Euclidean distances.}
\item{\code{similarity.scores}}{A dataframe of similarity scores derive from \code{distance.matrix}. See Szmrecsanyi et al. (2019) for details.}
}
}
\description{
Calculate the second line of evidence for the VADIS method
}
\details{
The function loops through a list of model objects, extracts the coefficient estimates, and compiles them in a single dataframe.

For scaling, there are four options. The default, \code{"abs"} (absolute), scales by a constant term based on the maximum reasonable distance, and values are bounded between 0 and 1 (see Szmrecsanyi et al. 2019). \code{"minmax"} uses minmax normalization, defined as

\deqn{ x' = \frac{x - min(x)}{max(x) - min(x)}}{x' = (x - min(x))/(max(x) - min(x))}

Minmax scaling bound values between 0 and 1. \code{"mean"} uses mean normalization, defined as
\deqn{ x' = \frac{x - mean(x)}{max(x) - min(x)}}{x' = (x - mean(x))/(max(x) - min(x))}

If \code{scale = "none"} no scaling is applied.
}
\examples{
\dontrun{
data_list <- split(particle_verbs_short, particle_verbs_short$Variety, drop = TRUE)

fmla <- Response ~ DirObjWordLength + DirObjDefiniteness + DirObjGivenness + DirObjConcreteness + DirObjThematicity + DirectionalPP + PrimeType + Semantics + Surprisal.P + Surprisal.V + Register

glm_func <- function(x) glm(fmla, data = x, family = binomial)

glm_list <- lapply(data_list, glm_func)
names(glm_list) <- names(data_list)

line2 <- vadis_line2(glm_list, path = FALSE)
}
}
\references{
Szmrecsanyi, Benedikt, Jason Grafmiller & Laura Rosseel. 2019. Variation-Based Distance and Similarity Modeling: A Case Study in World Englishes. \emph{Frontiers in Artificial Intelligence} 2. \url{https://doi.org/10.3389/frai.2019.00023}.
}
\author{
Jason Grafmiller
}
