% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vadis_line3.R
\name{vadis_line3}
\alias{vadis_line3}
\title{Calculate the third line of evidence for the VADIS method}
\usage{
vadis_line3(mod_list, path = NULL, conditional = FALSE)
}
\arguments{
\item{mod_list}{A list of random forest model objects.}

\item{path}{Path in which to save the output as an R data file (\code{.rds}). If \code{NULL}, defaults to the current working directory. Set \code{path = FALSE} if you do not wish to save to file.}

\item{conditional}{logical. Should unconditional (default) or conditional permutation variable importance be computed. Only applies to \code{RandomForest-class} models from the \code{\link[party]{party}} package.}
}
\value{
A list of length 3
}
\description{
Calculate the third line of evidence for the VADIS method
}
\details{
The function loops through a list of model objects, extracts the coefficient estimates, and compiles them in a single dataframe.
}
\examples{
\dontrun{
lm_fnc <- function(x) lm(Sepal.Length ~ Petal.Length + Petal.Width, data = x)
rm_list <- fit.vadis.RM(iris, split.by = "Species", fit.func = lm_fnc,
  path = FALSE)
summary(rm_list[[1]])

line3 <- calc_line3(rm_list, path = FALSE)
}
}
\author{
Jason Grafmiller
}
